Переносит константы из файла mock/events.js в const.js
