+++ src/const.js
@@ -49,4 +49,21 @@ const Statistic = {
   TIME_SPEND: `time-spend`,
 };
 
-export {EURO_SYMBOL, MenuName, menuNames, MILLISECONDS_PER_DAY, SortType, FilterType, EmptyEvent, Mode, Statistic};
+const GroupType = {
+  TRANSFER: `Transfer`,
+  ACTIVITY: `Activity`,
+};
+const groupTypeToPreposition = new Map([
+  [GroupType.TRANSFER, `to`],
+  [GroupType.ACTIVITY, `in`],
+]);
+const groupToTypes = new Map([
+  [GroupType.TRANSFER, [`taxi`, `bus`, `train`, `ship`, `transport`, `drive`, `flight`]],
+  [GroupType.ACTIVITY, [`check-in`, `restaurant`, `sightseeing`]],
+]);
+const typeToGroup = new Map();
+groupToTypes.forEach((value, key) => value.forEach((type) => typeToGroup.set(type, key)));
+
+export {EURO_SYMBOL, MenuName, menuNames, MILLISECONDS_PER_DAY,
+  SortType, FilterType, EmptyEvent, Mode, Statistic,
+  GroupType, groupToTypes, groupTypeToPreposition, typeToGroup};
